capstone-cimb/services/api-python/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
prometheus-fastapi-instrumentator